ml_in_prod/hw1/requirements.txt
hydra-core==1.1.2
hydra_colorlog==1.1.0
marshmallow-dataclass==8.5.8
numpy==1.22.3
omegaconf==2.1.2
pandas==1.4.1
pandas_profiling==3.2.0
pytest==7.1.2
pyyaml==6.0
scikit-learn==1.0.2
